+++ tests/test_plt.py
@@ -1,7 +1,8 @@
 import numpy as np
 import pytest
 
-from nmrsim.plt import add_lorentzians, mplplot, mplplot_stick, mplplot_lineshape
+from nmrsim.plt import mplplot, mplplot_stick, mplplot_lineshape
+from nmrsim.math import add_lorentzians
 from tests.accepted_data import ADD_SIGNALS_DATASET
 from tests.dnmr_standards import TWOSPIN_SLOW